convert int to float distances in MapPoint.cc

+++ src/MapPoint.cc
@@ -382,13 +382,13 @@ void MapPoint::ComputeDistinctiveDescriptors()
     }
 
     // Take the descriptor with least median distance to the rest
-    int BestMedian = INT_MAX;
+    float BestMedian = INT_MAX;
     int BestIdx = 0;
     for(size_t i=0;i<N;i++)
     {
-        vector<int> vDists(Distances[i],Distances[i]+N);
+        vector<float> vDists(Distances[i],Distances[i]+N);
         sort(vDists.begin(),vDists.end());
-        int median = vDists[0.5*(N-1)];
+        float median = vDists[0.5*(N-1)];
 
         if(median<BestMedian)
         {
